skin

+++ Backend/makeupp/src/main/java/com/makeupp/makeupp/controller/userController.java
@@ -1,18 +1,15 @@
 package com.makeupp.makeupp.controller;
 
-import org.springframework.web.bind.annotation.RestController;
+import com.makeupp.makeupp.DTO.loginDTO;
 import com.makeupp.makeupp.DTO.responseDTO;
 import com.makeupp.makeupp.DTO.userDTO;
+import com.makeupp.makeupp.model.user;
 import com.makeupp.makeupp.service.userService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/v1/user")
@@ -36,15 +33,38 @@ public class userController {
     @GetMapping("/{id}")
     public ResponseEntity<Object> getOneUser(@PathVariable int id) {
         var usuario = userService.findById(id);
-        if (!usuario.isPresent()) 
+        if (!usuario.isPresent())
             return new ResponseEntity<>("", HttpStatus.NOT_FOUND);
         return new ResponseEntity<>(usuario, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Object> deleteUser(@PathVariable int id) {
-        var respuesta = userService.deleteUser(id);
+        if (!userService.findById(id).isPresent()) {
+            return new ResponseEntity<>(new responseDTO(HttpStatus.NOT_FOUND.toString(), "Usuario no encontrado"),
+                    HttpStatus.NOT_FOUND);
+        }
+
+        responseDTO respuesta = userService.deleteUser(id);
         return new ResponseEntity<>(respuesta, HttpStatus.OK);
     }
 
+    @PutMapping("/{id}")
+    public ResponseEntity<Object> updateUser(@PathVariable int id, @RequestBody userDTO user) {
+        responseDTO respuesta = userService.update(id, user);
+        if (respuesta.getStatus().equals(HttpStatus.NOT_FOUND.toString())) {
+            return new ResponseEntity<>(respuesta, HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(respuesta, HttpStatus.OK);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<?> login(@RequestBody loginDTO loginDTO) {
+        try {
+            user user = userService.login(loginDTO.getEmail(), loginDTO.getPassword());
+            return ResponseEntity.ok(user);
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Credenciales incorrectas");
+        }
+    }
 }
